requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Flask-Mail>=0.9.1
Werkzeug>=3.0.1
python-dotenv>=1.0.0
requests>=2.31.0
google-genai>=0.2.0
pydantic>=2.0.0
